- Added pyrpmdb2rpmdb tool to convert our database to a rpm database - Modified test9 install testapp to support profiling - Use RpmResolver operation definitions for operations - Finished the freshenPkgs() method in RpmController to select and filter the   right packages - Delete unused variables, especially larger ones (large lists etc) - Remove duplicate operations from operations list. Might happen due to updates   and concurrent obsoletes - Some more code cleanup in RpmController - For updates always select the newest package given - Massively clean up the CPIO class and remove unused code - Fix runScript() to behave exactly like rpm - Set file mods on already installed directories, too (like rpm does) - Use os.lchown() for symlinks, otherwise the targe will be chowned :) - Fix my pkgCompare() functions to convert epochs always to strings - Dropped the skipsig parameter in the RpmStreamIO classes - Added an optional parameter to read() for RpmStreamIO to skip the parsing of   the complete header and move directly to the cpio archive - Optimized the generateTag() method, one of our most time consuming methods - Implemented the RpmDB class very similar to our RpmPyDB class - Dropped the "filenames" from RpmPyDB as we can generate that list faster   internally and saving it takes ages - Implemented the RpmUserCache class and use it per package for installation,   safes about 10 Minutes for a complete FC3 install - Modified RpmPackage to actually only read a rpm header once and otherwise   always skip parsing it - If a post script fails we don't abort the operation anymore just print the   error message as all files have already been written/erased anyway - Drop some unecessary time consuming method calls in RpmPackage - First rough version of the '#' markers during operation - Fixed a bug in the verfifyInstallFile() method in RpmPackage where an config   file update wasn't detected properly

+++ pyrpm/functions.py
@@ -29,15 +29,14 @@ def runScript(prog=None, script=None, arg1=None, arg2=None):
         prog = "/bin/sh"
     if not os.path.exists("/var/tmp"):
         os.makedirs("/var/tmp")
-    (fd, tmpfilename) = tempfile.mkstemp(dir="/var/tmp/", prefix="rpm-tmp.")
-    if fd == None:
-        return 0
-    if script != None:
-        os.write(fd, script)
-    os.close(fd)
-    fd = None
     args = [prog]
-    if prog != "/sbin/ldconfig":
+    if script != None:
+        (fd, tmpfilename) = tempfile.mkstemp(dir="/var/tmp/", prefix="rpm-tmp.")
+        if fd == None:
+            return 0
+        os.write(fd, script)
+        os.close(fd)
+        fd = None
         args.append(tmpfilename)
         if arg1 != None:
             args.append(arg1)
@@ -50,7 +49,8 @@ def runScript(prog=None, script=None, arg1=None, arg2=None):
         os.close(0)
         os.execv(prog, args)
         sys.exit()
-    os.unlink(tmpfilename)
+    if script != None:
+        os.unlink(tmpfilename)
     if status != 0:
         printError("Error in running script:")
         printError(str(prog))
@@ -77,6 +77,8 @@ def installFile(rfi, data):
             return 0
     elif filetype == CP_IFDIR:
         if os.path.isdir(rfi.filename):
+            if not setFileMods(rfi.filename, rfi.uid, rfi.gid, rfi.mode, rfi.mtime):
+                return 0
             return 1
         os.makedirs(rfi.filename)
         if not setFileMods(rfi.filename, rfi.uid, rfi.gid, rfi.mode, rfi.mtime):
@@ -91,6 +93,8 @@ def installFile(rfi, data):
         except:
             pass
         os.symlink(symlinkfile, rfi.filename)
+        if os.lchown(rfi.filename, rfi.uid, rfi.gid) != None:
+            return 0
     elif filetype == CP_IFIFO:
         makeDirs(rfi.filename)
         if not os.path.exists(rfi.filename) and os.mkfifo(rfi.filename) != None:
@@ -110,7 +114,10 @@ def installFile(rfi, data):
     return 1
 
 def setFileMods(filename, uid, gid, mode, mtime):
-    if os.chown(filename, uid, gid) != None:
+    try:
+        if os.chown(filename, uid, gid) != None:
+            return 0
+    except:
         return 0
     if os.chmod(filename, (~CP_IFMT) & mode) != None:
         return 0
@@ -326,14 +333,14 @@ def evrString(epoch, version, release):
 
 # Compare two packages by evr
 def pkgCompare(p1, p2):
-   if p1["epoch"] == None:
+    if p1["epoch"] == None:
         e1 = "0"
-   else:
-        e1 = p1["epoch"][0]
-   if p2["epoch"] == None:
+    else:
+        e1 = str(p1["epoch"][0])
+    if p2["epoch"] == None:
         e2 = "0"
-   else:
-        e2 = p2["epoch"][0]
-   return labelCompare((e1, p1["version"], p1["release"]), (e2, p2["version"], p2["release"]))
+    else:
+        e2 = str(p2["epoch"][0])
+    return labelCompare((e1, p1["version"], p1["release"]), (e2, p2["version"], p2["release"]))
 
 # vim:ts=4:sw=4:showmatch:expandtab
